Mark gemm nodes as unsafe
With the future (i.e. today or tomorrow) design of the multithreading
of this library, aliasing of matrices will occur, and data races will be
possible if the control tree is not set up correctly. Thus we mark gemm
nodes as unsafe as a lint.

As a very simple example, if there is a control node that opens threads,
and no control node underneath it that parallelizes ranges, then a data
race will occur as each thread will try to do exactly the same thing
when updating C.