Fixing bug when no barrier during GEMM
If there was no barrier, for smaller problems one thread would do its work
and then exit the parallel region, and then the same thread (with the same thread id)
would do the work of another thread.

This was an issue because of the cached data used by the thread would be wrong
when doing another thread's work.

The barrier inserted after computing the dgemm *apparently* fixes this problem.
However it's possible that this doesn't always work.

A permanent solution is to remove the caching mechanism
(each thread has a stateful control tree, and it is cached).
However then each thread has to create a new control tree every time
and this is incredibly slow right now.