Initial work on overlapping packing & communication
Introduced PackPair matrix type.
It's a wrapper around two matrices,
one of which is to be packed into the other

Added establish_leaf function to RoCM trait.
for PackPair matrices this will pack an RoCM leaf.
For other RoCM matrices it's a noop,
and kernels will have to call establish_leaf before
operating on raw RoCM matrices