Support for BLIS auxinfo
Use bindgen to generate bindings to blis typedefs
This lets us interface with the microkernel in a non-hacky way,
Before, we didn't pass in the auxinfo, and if the microkernel
accessed it, it would crash.
Now, the auxinfo is passed in properly,
however a_next and b_next are not appropriately set for performance.